requirements.txt
PyQt5==5.15.9
Pillow==9.5.0
wmi==1.5.1